Move schedules.js to branch worker
