+++ lib/workers/package-file/index.js
@@ -1,6 +1,6 @@
 const configParser = require('../../config');
 const depTypeWorker = require('../dep-type');
-const schedule = require('../package/schedule');
+const schedule = require('../branch/schedule');
 
 let logger = require('../../logger');
