+++ lib/workers/package/index.js
@@ -1,6 +1,6 @@
 const npmApi = require('../../api/npm');
 const versions = require('./versions');
-const schedule = require('./schedule');
+const schedule = require('../branch/schedule');
 const configParser = require('../../config');
 
 let logger = require('../../logger');
